# Speed up `has_path` by avoiding `popfirst!` and `push!`