@simonschoelly 2024-11-20
@thchr Can you rebase your PR on the lastest master? I think the nightly tests might succeed then.