@gdalle 2024-11-21
We could use an actual queue from DataStructures.jl instead of a vector though